+++ Wishboard/domain/Models/APIModels/AuthenticationModel/LostPasswordModel/LostPasswordModel.swift
@@ -1,0 +1,10 @@
+//
+//  LostPasswordModel.swift
+//  Wishboard
+//
+//  Created by gomin on 2022/09/28.
+//
+
+struct LostPasswordModel: Decodable {
+    let verificationCode: String?
+}
